requirements.txt
streamlit
openai
extract-msg
olefile
pandas
xlsxwriter
python-dateutil
